Move url parsing to a separate file
