+++ src/sections/Shard/Carousel.tsx
@@ -5,25 +5,12 @@ import { BsChevronLeft, BsChevronRight } from 'react-icons/bs';
 import { AnimatePresence, motion } from 'framer-motion';
 import { DateTime } from 'luxon';
 import { useHeaderFx } from '../../context/HeaderFx';
-import { getShardInfo, findNextShard } from '../../shardPredictor';
+import { getShardInfo } from '../../shardPredictor';
+import { parseUrl, replaceUrl } from '../../utils/parseUrl';
 import { ShardMapInfographic, ShardDataInfographic } from './Infographic';
 import ShardSummary from './Summary';
 
 const appZone = 'America/Los_Angeles';
-
-const relDateMap = {
-  eytd: -2,
-  ereyesterday: -2,
-  ytd: -1,
-  yesterday: -1,
-  tmr: 1,
-  tomorrow: 1,
-  ovmr: 2,
-  overmorrow: 2,
-};
-
-const replaceUrl = (path: string, pushHistory = true, state: unknown = null) =>
-  pushHistory ? history.pushState(state, '', path) : history.replaceState(state, '', path);
 
 const roundToRefDate = (date: DateTime, ref: DateTime) =>
   date.hasSame(ref, 'day') ? ref : date < ref ? date.endOf('day') : date.startOf('day');
@@ -43,48 +30,11 @@ const varients = {
   }),
 };
 
-const getDateFromUrl = () => {
-  const today = DateTime.local().setZone('America/Los_Angeles').startOf('day');
-  const path = window.location.pathname;
-  if (path === '/') return today;
-  const [route, ...params] = path.split('/').slice(1);
-  if (route === 'date') {
-    const [yearStr, monthStr, dayStr] = params;
-    const year = parseInt(yearStr.length === 2 ? `20${yearStr}` : yearStr, 10);
-    const month = monthStr ? parseInt(monthStr, 10) : 1;
-    const day = dayStr ? parseInt(dayStr, 10) : 1;
-
-    if (year && month && day) {
-      const date = DateTime.local(year, month, day, { zone: appZone });
-      if (date.isValid) {
-        if (date < DateTime.local(2022, 10, 1, { zone: appZone })) {
-          return DateTime.local(2022, 10, 1, { zone: appZone });
-        } else if (!DateTime.local({ zone: appZone }).hasSame(date, 'day')) {
-          return roundToRefDate(date, today);
-        }
-      }
-    }
-  } else if (Object.keys(relDateMap).includes(route)) {
-    const offset = relDateMap[route as keyof typeof relDateMap];
-    const date = today.plus({ days: offset });
-    replaceUrl(`/date/${date.toFormat('yyyy/MM/dd')}`, false);
-    return roundToRefDate(date, today);
-  } else if (route === 'next') {
-    const today = DateTime.local().setZone('America/Los_Angeles');
-    const color = params[0] as 'red' | 'black' | undefined;
-    const date = findNextShard(today, color && { only: color }).date;
-    replaceUrl(`/date/${date.toFormat('yyyy/MM/dd')}`, false);
-    return roundToRefDate(date, today);
-  }
-  replaceUrl('/', false);
-  return today;
-};
-
 export default function ShardCarousel() {
   const { t } = useTranslation(['shardCarousel', 'shard']);
 
   const [direction, setDirection] = useState(0);
-  const [date, setDate] = useState(() => getDateFromUrl());
+  const [date, setDate] = useState(() => roundToRefDate(parseUrl(), DateTime.local().setZone(appZone).startOf('day')));
   const info = useMemo(() => getShardInfo(date), [date.day, date.month, date.year]);
 
   const navigateDate = useCallback(
@@ -113,7 +63,7 @@ export default function ShardCarousel() {
   useEffect(() => {
     let timeout: string | number | NodeJS.Timeout | undefined = undefined;
     const handleDateChange = () => {
-      const newDate = getDateFromUrl();
+      const newDate = roundToRefDate(parseUrl(), DateTime.local().setZone(appZone).startOf('day'));
       if (!newDate.hasSame(date, 'day')) {
         navigateDate(newDate, false);
       }
